Refactor methods and attributes into Time class
